add concerns and refactor code
